interact with the contract when user is connected to fetch the gm messages
